emitters: store bf16/fp8 buffers at native stride
bf16/fp8 GEMM read garbage past element 0 on every GPU backend: the
host uploads tight narrow data (Field<u16>/Field<u8>, the layout the
CPU executor reads) while the MSL/SPIR-V emitters addressed one element
per u32 word. Fix the emitters to the host's contract — native stride
per backend; WGSL keeps its u32-slot with a documented host repack.

MSL (JIT + AOT): bf16 buffers are `ushort`, fp8 buffers are `uchar`.
Loads widen to uint before the unpack; stores narrow the packed bits
into the slot. The AOT emitter also gains the fp8 helper preamble and
the bf16/fp8/int4 load-store arms it was missing entirely.

SPIR-V (JIT + AOT): bf16 storage is a 16-bit int under
StorageBuffer16BitAccess (core in SPIR-V 1.3); fp8 is an 8-bit int
under StorageBuffer8BitAccess + OpExtension SPV_KHR_8bit_storage.
ArrayStride 2/1, aligned loads/stores, OpUConvert widen/narrow at the
boundary. Push-constant members keep the u32 slot so narrow scalars
don't drag in storagePushConstant16/8. The EmitCaps native-storage fork
is gone — native stride is the storage contract, not a device option.

Vulkan driver: query storageBuffer16BitAccess / storageBuffer8BitAccess
via vkGetPhysicalDeviceFeatures2 on 1.2+ devices and enable them at
vkCreateDevice; the caps table marks bf16/fp8 as feature-gated.

WGSL cannot hold 16-/8-bit storage elements, so it stays u32-slot. The
new Gpu::narrow_storage_u32_slot() reports that layout and the
mixed-GEMM dispatch re-uploads A/B expanded one-element-per-word on
such a backend, keeping the public tight-Field contract uniform.

The gemm_mixed tests are un-pinned from the CPU JIT and run on the real
GPU under gpu-metal/gpu-vulkan: bf16 + fp8 matrix/identity/gemv all
pass on Metal. A new emit_spirv_narrow_storage test pins the
capability/stride contract and runs spirv-val over gemm-shaped bf16/fp8
modules; op-matrix diff lanes stay green (single-element u32 uploads
read identically under either layout).

+++ crates/quanta-ir/src/types.rs
@@ -730,13 +730,16 @@ impl ScalarType {
         }
     }
 
-    /// MSL *buffer storage* element type. Differs from `msl_name` only for
-    /// bf16/fp8 (one per `uint` word) and int4 (8 nibbles per `uint` word):
-    /// pack/unpack happens at load/store.
+    /// MSL *buffer storage* element type. Differs from `msl_name` for
+    /// bf16 (`ushort`, native 2-byte stride), fp8 (`uchar`, native 1-byte
+    /// stride) and int4 (8 nibbles per `uint` word): pack/unpack happens at
+    /// load/store. The narrow-float strides match the host upload layout
+    /// (tight `Field<u16>` / `Field<u8>`) and the CPU executor; only WGSL
+    /// keeps the u32-slot fallback (see `wgsl_storage_name`).
     pub fn msl_storage_name(&self) -> &'static str {
         match self {
-            Self::BF16 => "uint",
-            Self::FP8E5M2 | Self::FP8E4M3 => "uint",
+            Self::BF16 => "ushort",
+            Self::FP8E5M2 | Self::FP8E4M3 => "uchar",
             Self::I4 => "uint",
             _ => self.msl_name(),
         }
@@ -765,6 +768,12 @@ impl ScalarType {
     /// WGSL *buffer storage* element type. Differs from `wgsl_name` for
     /// bf16/fp8 (one per `u32` word) and int4 (8 nibbles per `u32` word):
     /// pack/unpack at load/store.
+    ///
+    /// WGSL storage buffers cannot hold 16-/8-bit array elements, so bf16
+    /// and fp8 keep the u32-slot layout here even though MSL/SPIR-V (and
+    /// the host/CPU) use native stride. Hosts feeding the WebGPU backend
+    /// must expand tight narrow data one-element-per-word (see
+    /// `Gpu::narrow_storage_u32_slot`).
     pub fn wgsl_storage_name(&self) -> &'static str {
         match self {
             Self::BF16 => "u32",
